+++ geocoder/src/iosMain/kotlin/tech/kotlinlang/geocoder/GeocoderHelper.ios.kt
@@ -1,0 +1,5 @@
+package tech.kotlinlang.geocoder
+
+actual fun getGeocoderHelper(): GeocoderHelper {
+    return IosGeocoderHelper()
+}