+++ permission/src/androidMain/kotlin/tech/kotlinlang/permission/location/AndroidLocationHelper.kt
@@ -53,8 +53,4 @@ class AndroidLocationHelper(
             }
         }
     }
-
-    override suspend fun fetchCurrentLocation(): LocationRequestResult {
-        TODO("Not yet implemented")
-    }
 }